// src/shared/Enums/Stylings.tsx
export enum Colors {
  MAIN = "main ",
  SECONDARY = "secondary ",
  SECONDARY_LIGHT = "secondarylight ",
  THIRD = "third ",
  FOURTH = "fourth ",
  FIFTH = "fifth ",
  WHITE = "white ",
}

export enum Sizes {
  XS = "xs ",
  SM = "sm ",
  BASE = "base ",
  LG = "lg ",
  XL = "xl ",
  XL2 = "2xl ",
  XL3 = "3xl ",
}